requirements.txt
streamlit>=1.31.1
python-dotenv>=1.0.0
openai>=1.12.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
more-itertools>=10.1.0
openpyxl
pandas>=2.0.0
more-itertools
google-generativeai>=0.3.0
numpy>=1.24.0; python_version >= '3.9' and platform_machine != 'arm64' and platform_system != 'Windows' and platform_machine != 'aarch64'
numpy>=1.24.0; python_version >= '3.9' and platform_machine == 'x86_64' and platform_system == 'Windows'
--find-links https://download.pytorch.org/whl/torch_stable.html